calendar cells now highlighting based on api call
