+++ src/app.js
@@ -1,6 +1,5 @@
 import React from "react";
 import Calendar from "./components/Calendar";
-import Registrations from "./components/Registrations";
 import "./styles/App.css";
 
 class App extends React.Component {
@@ -17,7 +16,6 @@ class App extends React.Component {
         <main>
           <Calendar />
         </main>
-        <Registrations />
       </div>
     );
   }
