+++ src/components/Calendar.js
@@ -6,8 +6,20 @@ class Calendar extends React.Component {
   state = {
     currentMonth: new Date(2025, 7),
     selectedDate: new Date(2025, 7, 1),
-    registrationDays: { dates: [new Date(2025, 7, 5), new Date(2025, 7, 6)]},
+    registrationDays: {dates: []},
   };
+
+  componentDidMount() {
+    fetch('https://demo14.secure.retreat.guru/api/v1/registrations?token=ef061e1a717568ee5ca5c76a94cf5842')
+      .then(response => response.json())
+      .then(data => {
+        let dates = [1];
+        let occupiedDates = data.map((reg) => {
+          dates = dates.concat(dateFns.eachDayOfInterval({ start: dateFns.parse(reg.start_date, 'yyyy-MM-dd', new Date()), end: dateFns.parse(reg.end_date, 'yyyy-MM-dd', new Date()) }));
+        })
+        this.setState({registrationDays: {dates: dates}});
+      })
+  }
 
   renderHeader() { 
     const dateFormat = "MMMM yyyy";
